Working MSI building

+++ scripts/windows-full-deploy/1.build-and-prepare-windows-binary.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
-cd ..
-./windows-build.sh
+FILE=LICENSE
+
+if [ ! -f "$FILE" ]; then
+    echo "Please run script from main directory"
+    exit 1
+fi
+
+./scripts/windows-build.sh
 mkdir -p generated/builds/windows/
 cp "src/deploy/windows/src.exe" "generated/builds/windows/ThreeFoldNetworkConnector.exe"
 cp libs/wintun.dll generated/builds/windows/wintun.dll
